+++ app/schemas/transaction.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, ConfigDict
@@ -8,14 +9,14 @@ from app.core.enums import CurrencyEnum, TransactionStatusEnum
 
 class RequestTransactionModel(BaseModel):
     currency: CurrencyEnum
-    amount: float
+    amount: Decimal
 
 
 class TransactionModel(BaseModel):
     id: Optional[int]
     user_id: Optional[int] = None
     currency: Optional[CurrencyEnum] = None
-    amount: Optional[float] = None
+    amount: Optional[Decimal] = None
     status: Optional[TransactionStatusEnum] = None
     created: Optional[datetime] = None
